Add default layout configuration
For now this is a simple string entered through a LineEdit.

The format is:

desktop:[layoutname][,desktop:layoutname]

where desktop is the number of the desktop like the user sees
it (i.e. starting with 1)

A negative desktop means tiling is disabled by default for that desktop.

A zero desktop changes the default layout for all desktops not
explicitly set to something else.

I'm not sure how to document this yet (a "Usage" section in README.md?)