Delay tiling when compositing is on again
When we tile too early, a client might show up broken (mostly this
manifests as e.g. dolphin not being shown at all until it is resized).

Unfortunately, this breaks when compositing is enabled/disabled after
the script is started. However, that's a KWin bug and should be fixed there.

Sorry!