Basic implementation of I3Layout
The way I decided to implement this layout differs a bit from the
others. In the original layouts, the main data structure is the
tiles array, and the position of a tile in this array would define
its position.

In I3Layout, the position is managed inside a containerTree class,
and the tiles array is only kept to interface with the original code.
This means the newly added tiles must always be pushed at the end, and
calling I3Layout.addTile(x,y) will define the last tile as a newly
allocated tree node. The allocation algorithm (still not implemented)
will simply select the tile at position x,y and the new tile will be
allocated at the same container.

When creating new windows, the x,y position will be set as the currently
focused window's position. This will create an intuitive behaviour where
the newly added window is spawned next/below the currently focused
tile's container. Passing down the currently (actually previous, for
implementation reasons) focused windows required some breaking changes
in the Tiling.addTile API.