Detect focused window && basic allocation algorithm
Fixed some things in the detection of focused windows. For example,
when creating a new client, its focused (active) state will be set
before appearing at the clients list. To fix this, I set the focused
flag on clientAdded as well.

The basic allocation algorithm is implemented. When a new tile must be
created inside an I3Layout, the most suitable location is selected and
a new tile is allocated there. This is now tested for a single horizontal
container.