support window maximization
Currently, maximizing a window doesn't do anything in tiling mode. This
commit minimizes all but the current window if a maximize operation
(such as double-clicking the title bar or via binding "KWin/Maximize
Window" to a global keyboard shortcut) is requested by the user.

Credits of the function minimizeOthers goes to
https://github.com/karakum/kwin-script-onlyOneActiveWindow