Fix missed geometry change events
KWin sometimes decides to update geometry of several/all clients
simultaneously. Our timer-based asynchronous handler needs to be
ready for that. It also needs to handle geometry changes happening
to clients outside of the current desktop, because some effects
such as Desktop Grid display badly damaged layouts if it doesn't.